+++ tests/test_changes.py
@@ -1,4 +1,8 @@
+import mock
+import time
 import unittest
+
+from kinto.core import utils as core_utils
 
 from kinto_changes import __version__ as changes_version
 from . import BaseWebTest, get_user_headers
@@ -13,43 +17,40 @@ class UpdateChangesTest(BaseWebTest, unittest.TestCase):
 
     def setUp(self):
         super(UpdateChangesTest, self).setUp()
-        # XXX: should happen during Listener instanciation or includeme()
-        self.app.post_json(self.records_uri,
-                           SAMPLE_RECORD,
+        self.app.post_json(self.records_uri, SAMPLE_RECORD,
                            headers=self.headers)
+
+    def test_parent_bucket_and_collection_dont_have_to_exist(self):
+        self.app.delete('/buckets/monitor/collections/changes',
+                        headers=self.headers, status=(403, 404))
+        self.app.get(self.changes_uri)  # Not failing
+        self.app.delete('/buckets/monitor',
+                        headers=self.headers, status=(403, 404))
+        self.app.get(self.changes_uri)  # Not failing
+
+    def test_parent_bucket_and_collection_can_exist(self):
+        self.app.put('/buckets/monitor', headers=self.headers)
+        resp = self.app.get(self.changes_uri)  # Not failing
+        self.assertEqual(len(resp.json['data']), 1)
+
+        self.app.put('/buckets/monitor/collections/changes', headers=self.headers)
+        resp = self.app.get(self.changes_uri)  # Not failing
+        self.assertEqual(len(resp.json['data']), 1)
 
     def test_a_change_record_is_updated_per_bucket_collection(self):
-        resp = self.app.get(self.changes_uri, headers=self.headers)
-        change_record = resp.json['data'][0]
+        resp = self.app.get(self.changes_uri)
+        before_timestamp = resp.json['data'][0]['last_modified']
+        before_id = resp.json['data'][0]['id']
 
         self.app.post_json(self.records_uri, SAMPLE_RECORD,
                            headers=self.headers)
 
-        resp = self.app.get(self.changes_uri, headers=self.headers)
-        after = resp.json['data'][0]
-        self.assertEqual(change_record['id'], after['id'])
-        self.assertNotEqual(change_record['last_modified'],
-                            after['last_modified'])
+        resp = self.app.get(self.changes_uri)
 
-    def test_collection_on_different_host_do_not_have_the_same_id(self):
-        self.app.post_json(self.records_uri, SAMPLE_RECORD,
-                           headers=self.headers)
-        self.app.app.registry.settings['http_host'] = 'another.host.com'
-        self.app.post_json(self.records_uri, SAMPLE_RECORD,
-                           headers=self.headers)
-        resp = self.app.get(self.changes_uri, headers=self.headers)
-        self.assertEquals(len(resp.json['data']), 2)
-
-    def test_changes_bucket_and_collection_are_created_automatically(self):
-        self.app.delete('/buckets/monitor', headers=self.headers)
-        self.app.post_json(self.records_uri, SAMPLE_RECORD,
-                           headers=self.headers)
-        self.app.get(self.changes_uri, headers=self.headers, status=200)
-
-    def test_changes_is_readable_by_everyone_by_default(self):
-        headers = self.headers.copy()
-        del headers['Authorization']
-        self.app.get(self.changes_uri, headers=headers, status=200)
+        after_timestamp = resp.json['data'][0]['last_modified']
+        after_id = resp.json['data'][0]['id']
+        self.assertEqual(before_id, after_id)
+        self.assertNotEqual(before_timestamp, after_timestamp)
 
     def test_only_collections_specified_in_settings_are_monitored(self):
         resp = self.app.get(self.changes_uri, headers=self.headers)
@@ -65,29 +66,25 @@ class UpdateChangesTest(BaseWebTest, unittest.TestCase):
         self.assertEqual(change_record['last_modified'],
                          after['last_modified'])
 
-    def test_changes_bucket_last_modified_is_not_updated(self):
-        resp = self.app.get('/buckets/monitor', headers=self.headers)
-        before = resp.json['data']['last_modified']
+    def test_the_resource_configured_can_be_a_collection_uri(self):
+        with mock.patch.dict(self.app.app.registry.settings,
+                             [('changes.resources',
+                               '/buckets/blocklists/collections/certificates')]):
+            resp = self.app.get(self.changes_uri)
+        self.assertEqual(len(resp.json['data']), 1)
 
-        self.app.post_json(self.records_uri, SAMPLE_RECORD,
-                           headers=self.headers)
+    def test_returns_304_if_no_change_occured(self):
+        resp = self.app.get(self.changes_uri)
+        before_timestamp = resp.headers["ETag"]
+        self.app.get(self.changes_uri,
+                     headers={'If-None-Match': before_timestamp},
+                     status=304)
 
-        resp = self.app.get('/buckets/monitor', headers=self.headers)
-        after = resp.json['data']['last_modified']
-        self.assertEqual(before, after)
-
-    def test_changes_collection_last_modified_is_not_updated(self):
-        resp = self.app.get('/buckets/monitor/collections/changes',
-                            headers=self.headers)
-        before = resp.json['data']['last_modified']
-
-        self.app.post_json(self.records_uri, SAMPLE_RECORD,
-                           headers=self.headers)
-
-        resp = self.app.get('/buckets/monitor/collections/changes',
-                            headers=self.headers)
-        after = resp.json['data']['last_modified']
-        self.assertEqual(before, after)
+    def test_returns_empty_list_if_no_resource_configured(self):
+        with mock.patch.dict(self.app.app.registry.settings,
+                             [('changes.resources', '')]):
+            resp = self.app.get(self.changes_uri)
+        self.assertEqual(resp.json['data'], [])
 
     def test_change_record_has_greater_last_modified_of_collection_of_records(self):
         resp = self.app.post_json(self.records_uri, SAMPLE_RECORD,
@@ -134,43 +131,8 @@ class UpdateChangesTest(BaseWebTest, unittest.TestCase):
             "description": "Track modifications of records in Kinto and store "
                            "the collection timestamps into a specific bucket "
                            "and collection.",
-            "collections": [u'/buckets/blocklists'],
+            "collections": ['/buckets/blocklists'],
             "url": "http://kinto.readthedocs.io/en/latest/tutorials/"
                    "synchronisation.html#polling-for-remote-changes"
         }
         self.assertEqual(expected, capabilities['changes'])
-
-
-class UpdateConfiguredChangesTest(BaseWebTest, unittest.TestCase):
-    config = 'mozilla.ini'
-
-    def setUp(self):
-        super(UpdateConfiguredChangesTest, self).setUp()
-        records_uri = '/buckets/default/collections/certificates/records'
-        self.app.post_json(records_uri,
-                           SAMPLE_RECORD,
-                           headers=self.headers)
-
-    def test_changes_bucket_and_collection_can_be_specified_in_settings(self):
-        # See mozilla.ini
-        changes_uri = '/buckets/mozilla/collections/updates/records'
-        resp = self.app.get(changes_uri, headers=self.headers)
-        self.assertEquals(len(resp.json['data']), 1)
-
-    def test_changes_records_permissions_can_be_specified_in_settings(self):
-        self.app.get('/buckets/mozilla/collections/updates',
-                     headers=self.headers,
-                     status=200)
-
-    def test_changes_records_are_not_readable_for_unauthorized(self):
-        headers = self.headers.copy()
-        headers.update(**get_user_headers('tartampion'))
-        self.app.get('/buckets/mozilla/collections/updates/records',
-                     headers=headers,
-                     status=403)
-
-    def test_change_record_has_specific_server_host(self):
-        changes_uri = '/buckets/mozilla/collections/updates/records'
-        resp = self.app.get(changes_uri, headers=self.headers)
-        self.assertEqual(resp.json['data'][0]['host'],
-                         'firefox.settings.services.mozilla.com')